+++ NEWS.md
@@ -3,6 +3,9 @@
 * Fixed issue in `check_data()` where missing responses resulted in an error
   when `missing` was specified.
 
+* `check_missing()` can now be used to ensure that a value is a length 1 scalar
+  (of any type).
+
 # rdcmchecks 0.1.0
 
 * Initial CRAN submission.